Implement expansion within meta predicate scope
The code is rewritten for the most part, because we can't just move
guards to the top level of the expansion. So instead we need to
accumulate guards and then either send them up to the parent term or
prepend them to the current term depending on the meta_predicate
specification of the calling goal.

To deal with meta arguments that require N>0 number of additional
arguments we construct a lambda term (as in library(yall)) to prepend
the resulting guard and then pipe the outside arguments to the expanded
term. I.e. if T expands to T' with guard G in the first argument of
call/3 we have an expansion of the form

```
call(T, X, Y) ==> call([A, B] >> (H, call(T', A, B)), X, Y)
```